+++ src/pages/employer/RemittanceManagePage.jsx
@@ -297,3 +297,4 @@ export default function RemittanceManagePage() {
     </div>
   );
 }
+
